- The connection page for bluetooth devices is now set to home page (java and xml changed for both) - The settings page will be implemented later - Changed java files names to have the first letter as caps
