+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Android\.android\avd\Pixel_4_API_33_1.avd" />
+            <value value="C:\Users\Harrisson\.android\avd\Pixel_3a_API_33_x86_64.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-11-14T23:25:58.147616700Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-11-18T03:28:13.864106100Z" />
   </component>
 </project>
