+++ app/src/main/java/com/example/cprfeedbackapp/HomeFragment.java
@@ -1,14 +1,28 @@
 package com.example.cprfeedbackapp;
 
-import android.content.DialogInterface;
+import android.Manifest;
+import android.bluetooth.BluetoothAdapter;
+import android.bluetooth.BluetoothDevice;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
 
+import androidx.core.app.ActivityCompat;
 import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.Toast;
+
+import com.google.android.material.snackbar.Snackbar;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -16,6 +30,14 @@ import android.widget.Button;
  * create an instance of this fragment.
  */
 public class HomeFragment extends Fragment {
+
+    private Button connectButton;
+    private Button scanButton;
+    private Button buttonGetData;
+
+    protected RecyclerView recyclerView;
+    protected List<DeviceInfoModel> deviceList;
+
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -25,7 +47,6 @@ public class HomeFragment extends Fragment {
     // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
-
 
     public HomeFragment() {
         // Required empty public constructor
@@ -37,7 +58,7 @@ public class HomeFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment HomeFragment.
+     * @return A new instance of fragment SettingsFragment.
      */
     // TODO: Rename and change types and number of parameters
     public static HomeFragment newInstance(String param1, String param2) {
@@ -58,13 +79,69 @@ public class HomeFragment extends Fragment {
         }
     }
 
+    @Override
+    public void onResume() {
+        super.onResume();
+
+
+    }
+
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-
-        View fragmentView = inflater.inflate(R.layout.fragment_home, container, false);
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
 
         // Inflate the layout for this fragment
+        View fragmentView = inflater.inflate(R.layout.fragment_home, container, false);
+
+        BluetoothServiceManager btServiceManager = new BluetoothServiceManager(HomeFragment.this.getContext(), HomeFragment.this.getActivity());
+
+        Set<BluetoothDevice> pairedDevices = btServiceManager.queryPairedDevice();
+
+        deviceList = new ArrayList<>();
+
+        if (pairedDevices != null){
+
+            if (pairedDevices.size() > 0) {
+                for (BluetoothDevice device : pairedDevices) {
+
+                    if (ActivityCompat.checkSelfPermission(HomeFragment.this.getContext(), Manifest.permission.BLUETOOTH_CONNECT) == PackageManager.PERMISSION_GRANTED) {
+
+                        String deviceName = device.getName();
+                        String deviceHardwareAddress = device.getAddress();
+                        DeviceInfoModel deviceInfoModel = new DeviceInfoModel(deviceName, deviceHardwareAddress);
+                        deviceList.add(deviceInfoModel);
+
+                    }
+                }
+
+                //recyclerView = fragmentView.findViewById(R.id.deviceRecyclerView);
+                recyclerView = fragmentView.findViewById(R.id.deviceRecyclerView);
+                setupRecyclerView();
+            }
+
+            else
+                Toast.makeText(HomeFragment.this.getContext(), "Pair a Bluetooth Device.", Toast.LENGTH_LONG).show();
+        }
+
+        else
+            Toast.makeText(HomeFragment.this.getContext(), "Can't Display a device", Toast.LENGTH_LONG).show();
+
         return fragmentView;
     }
+
+    public void setupRecyclerView() {
+
+        // Display paired device using recyclerView
+        recyclerView.setLayoutManager(new LinearLayoutManager(HomeFragment.this.getContext()));
+        DeviceRecyclerViewAdapter deviceListAdapter = new DeviceRecyclerViewAdapter(deviceList, HomeFragment.this.getContext());
+        recyclerView.setAdapter(deviceListAdapter);
+    }
+
+
+    //For toasts
+    private void msg(String str) {
+        Toast.makeText(HomeFragment.this.getContext(), str, Toast.LENGTH_LONG).show();
+    }
+
 }
