+++ app/src/main/java/com/example/cprfeedbackapp/SettingsFragment.java
@@ -1,28 +1,12 @@
 package com.example.cprfeedbackapp;
 
-import android.Manifest;
-import android.bluetooth.BluetoothAdapter;
-import android.bluetooth.BluetoothDevice;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 
-import androidx.core.app.ActivityCompat;
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.Toast;
-
-import com.google.android.material.snackbar.Snackbar;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -30,14 +14,6 @@ import java.util.Set;
  * create an instance of this fragment.
  */
 public class SettingsFragment extends Fragment {
-
-    private Button connectButton;
-    private Button scanButton;
-    private Button buttonGetData;
-
-    protected RecyclerView recyclerView;
-    protected List<DeviceInfoModel> deviceList;
-
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -58,7 +34,7 @@ public class SettingsFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment SettingsFragment.
+     * @return A new instance of fragment healthMonitoringFragment.
      */
     // TODO: Rename and change types and number of parameters
     public static SettingsFragment newInstance(String param1, String param2) {
@@ -80,98 +56,9 @@ public class SettingsFragment extends Fragment {
     }
 
     @Override
-    public void onResume() {
-        super.onResume();
-
-
-    }
-
-
-    @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-
         // Inflate the layout for this fragment
-        View fragmentView = inflater.inflate(R.layout.fragment_settings, container, false);
-//      BluetoothServiceManager bluetoothServiceManager = new BluetoothServiceManager(SettingsFragment.this.getContext(), SettingsFragment.this.getActivity());
-
-//        scanButton = fragmentView.findViewById(R.id.scanButton);
-//
-//        scanButton.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                bluetoothServiceManager.discoverDevices();
-//            }
-//        });
-//
-//        connectButton = fragmentView.findViewById(R.id.connectButton);
-//
-//        connectButton.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                bluetoothServiceManager.connectBluetoothDevice();
-//            }
-//        });
-//
-//        buttonGetData = fragmentView.findViewById(R.id.buttonDisplayData);
-//        buttonGetData.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                BluetoothServiceManager bt = new BluetoothServiceManager(SettingsFragment.this.getContext(), SettingsFragment.this.getActivity());
-//                if(bt.checkBluetoothEnabled())
-//                {
-//                    bt.getInputData();
-//                }
-//
-//            }
-//        });
-        BluetoothServiceManager btServiceManager = new BluetoothServiceManager(SettingsFragment.this.getContext(), SettingsFragment.this.getActivity());
-
-        Set<BluetoothDevice> pairedDevices = btServiceManager.queryPairedDevice();
-
-        deviceList = new ArrayList<>();
-
-        if (pairedDevices != null){
-
-            if (pairedDevices.size() > 0) {
-                for (BluetoothDevice device : pairedDevices) {
-
-                    if (ActivityCompat.checkSelfPermission(SettingsFragment.this.getContext(), Manifest.permission.BLUETOOTH_CONNECT) == PackageManager.PERMISSION_GRANTED) {
-
-                        String deviceName = device.getName();
-                        String deviceHardwareAddress = device.getAddress();
-                        DeviceInfoModel deviceInfoModel = new DeviceInfoModel(deviceName, deviceHardwareAddress);
-                        deviceList.add(deviceInfoModel);
-
-                    }
-                }
-
-                recyclerView = fragmentView.findViewById(R.id.deviceRecyclerView);
-                setupRecyclerView();
-            }
-
-            else
-                Toast.makeText(SettingsFragment.this.getContext(), "Pair a Bluetooth Device.", Toast.LENGTH_LONG).show();
-        }
-
-        else
-            Toast.makeText(SettingsFragment.this.getContext(), "Can't Display a device", Toast.LENGTH_LONG).show();
-
-        return fragmentView;
+        return inflater.inflate(R.layout.fragment_settings, container, false);
     }
-
-    public void setupRecyclerView() {
-
-        // Display paired device using recyclerView
-        recyclerView.setLayoutManager(new LinearLayoutManager(SettingsFragment.this.getContext()));
-        DeviceRecyclerViewAdapter deviceListAdapter = new DeviceRecyclerViewAdapter(deviceList, SettingsFragment.this.getContext());
-        recyclerView.setAdapter(deviceListAdapter);
-    }
-
-
-    //For toasts
-    private void msg(String str) {
-        Toast.makeText(SettingsFragment.this.getContext(), str, Toast.LENGTH_LONG).show();
-    }
-
 }
