+++ app/src/main/java/com/example/cprfeedbackapp/TutorialFragment.java
@@ -10,10 +10,10 @@ import android.view.ViewGroup;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link tutorialFragment#newInstance} factory method to
+ * Use the {@link TutorialFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class tutorialFragment extends Fragment {
+public class TutorialFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +24,7 @@ public class tutorialFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public tutorialFragment() {
+    public TutorialFragment() {
         // Required empty public constructor
     }
 
@@ -37,8 +37,8 @@ public class tutorialFragment extends Fragment {
      * @return A new instance of fragment tutorialFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static tutorialFragment newInstance(String param1, String param2) {
-        tutorialFragment fragment = new tutorialFragment();
+    public static TutorialFragment newInstance(String param1, String param2) {
+        TutorialFragment fragment = new TutorialFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
