+++ app/src/main/res/layout/activity_data.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".dataActivity">
+    tools:context=".DataActivity">
 
     <TextView
         android:id="@+id/textView11"
