+++ app/src/main/res/layout/fragment_tutorial.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".tutorialFragment">
+    tools:context=".TutorialFragment">
 
     <TextView
         android:id="@+id/textView9"
